Convert python2 files to python3 with 2to3
This commit converts python2 files to python3 with 2to3.
It also tries to fixes resulting errors caused by that transition.

At the moment there seems to be a json conversion problem with
the sanitizer-expert, I don't know if that was always present or
if that was caused by the transition.

The error is:

2015-05-05 21:50:21,935 - sanitizer-expert - ERROR - Check the following exception:
Traceback (most recent call last):
  File "/usr/lib/python3.4/site-packages/intelmq-0.0.9-py3.4.egg/intelmq/lib/bot.py", line 69, in start
    self.process()
  File "/usr/lib/python3.4/site-packages/intelmq-0.0.9-py3.4.egg/intelmq/bots/experts/sanitizer/sanitizer.py", line 35, in process
AttributeError: 'str' object has no attribute 'contains'

And was caused by a message like:
2015-05-05 21:50:21,935 - sanitizer-expert - INFO - Current Message(event): b"{'observation_time': '2015-05-05T18:35:48', 'source_time': '2015-05-05T00:00:00+00:00', 'reported_source_ip': '190.255.48.99', 'type': 'brute-force', 'feed': 'arbor', 'feed_url': 'http://atlas-public.ec2.arbor.net/public/ssh_attackers', 'source_ip': '190.255.48.99'}"