Fix of redis<->intelmq bug caused by python3 uses of strings
This commit seems after a couple of manual tests to fix the
issue described in 3dbdc0d88e4000ef75a6182ae23082912e54d5ab
which was caused by python2 uses 8byte str/unicodde vs. python3
uses text/bytes encoding.

Which caused that when encoding
intelmq messages(of class Event) for redis were displayed as the
python dict representation. And not the actual wanted unicode
JSON representation, which was then stored in redis and from there
it could only be incorrectly retrieved. As it was a python dict
representation it couldn't be parsed as JSON it was handled as string
and not as an intelmq event. Which then caused an exception of calling
the contains method, which isn't defined on strings, only on intelmq events.

Better unit tests would help for those cases.