Adds TestBot bot test interaction test
This commit adds a TestCase for the modified bot interface,
which was added to test a bot without much hassles.

The TestCase tests some basic functionality of a bot, like:
Can it log something? Can it process prepared test messages
and send it over a TestPipeline? Do ERRORs happen in its log?

What it doesn't represent is yet a well-refactored interface,
it could be made easier by having a special `BotTestCase' which
does all those things this TestCase does explicit in the background.