ENH: intelmqctl log parsing/reading simplified
The logs are read and parsed backwards, which is more efficient than reading and parsing the entire file from the beginning
The code to read the file is easier to read and understand now, however some code in utils reads the file backwards in chunks.
This change is currently not compatible with intelmq-manager as it wants to read the line number which is not read anymore as its not possible to get it by reading the file backwards. Yet a line number does not make sense in log files.


Former-commit-id: 9cd3205759f17182371919cf9a591e35eabc3f16