ENH: bot tests check class name and event fields
It is checked if the Bot class has a CamelCase name
and ends with CollectorBot, ParserBot etc.

It the bot is a parser or expert, it will be checked if the
event has some essential fields like feed.url or raw.
Same applies to collectors.

bot_id is 'test-bot' by default, but can be changed.

Signed-off-by: Sebastian Wagner <sebix@sebix.at>