Store the network address as CIDR
PostgreSQL supports a cidr type, so use that to store network addresses.
This commit removes the index on the network addresses because the btree
index PostgreSQL uses by default will not be used in the queries we want
to make (lookup a network based on ip-address). The index will have to
be added again in later commit.