Use the new DB interface in the certbund_contact bot
This involves using the new DB function which returns quite a bit more
information about the recipients and the desired notifications which
therefore cannot be simply stored in the messages in the set of email
addresses under "source.abuse_contact" and "destination.abuse_contact".
The information is now stored in the JSON structure under "extra", using
a bot-specific key to avoid conflicts with other bots.

The test cases are adapated accordingly.