intelmq_psql_initdb.py JSON instead of String
Modified the file in order to create database columns of type
json instead of string.
This enhances the processing speed of data which is stored in the
extras column.

ATTENTION: This is a schema-alteration of the EventDB.
If data already exists in the EventDB it might be required, that data is
converted from string to json.
Refer to: http://www.postgresql.org/docs/9.2/static/datatype-json.html

This change requires a PostgreSQL 9.2 or higher.