The XMPP Bot is now capable to connect to a rom on an XMPP Server.
This amends commit 1b0bef96d801d28dd075cbd9679e10d76aefa164, somehow two files
were not pushed...

It is now possible to:
 1) Connect to a server using username / password
 2) Receive messages which were send to the bot directly
 3) Receive messages which were send in a room
 All Messages are send into the MessageQueue without further
 processing.

the room must be configured in format:
 room@server.example.com

I added a __init__.py to be able to load the bot as module.

The bot was tested briefly with sleekxmpp 1.0-beta5.
It responded as expected to direct messages and messages in rooms.
 - The use of password-protected rooms was not tested.

TODOs:
 - Write Tests for this Bot
 - Test MUC with Password manually
 - Enance Error Messages when the Bot was misconfigured
 - Tidy Up Code