Complete the schema changes for separate manual/automatic tables
The changes to the schema and the stored procedures for separate
sets of tables for entries that are manually maintained and ones
that are automatically updated should be complete with this
changeset. This changeset includes:

  - the is_manual columns are removed because they're not needed anymore

  - organisation_manual and autonomous_system_manual are added, along
    with the necessary corresponding nm-relation tables and changes to
    existing such tables

  - New procedures to query the *_manual tables:
      notifications_for_ip_manual
      notifications_for_asn_manual
      notifications_for_fqdn_manual

This changeset is bigger than I would like, but splitting it into
more managable/reviewable changesets would take too much time at
the moment.