Simplify the queries on the notifications_for_-functions
The matched_contacts sub-queries in the notifications_for_-functions now
yield the organisation_id instead of the contact_id and the
organisation_id is used to lookup the organisation specific parameters
like the templates to use. The matched_contacts sub-queries involve the
organisation anyway so reusing the information which organisation a
contact belongs to makes the queries simpler.

A consequence of this change is that the contact_organisation_settings
and contact_organisation_settings_automatic views do not involve
contacts anymore so they are renamed accordingly.