Match on classification.type instead of classification.identifier
The classification.type is much more useful at this stage because its
value is always set for the feeds were currently primarily interested
in. The classification.identifier is not always set, so we cannot rely
on it as the only classification criterion.

We're planning to allow more fine-grained criteria for notifications
beyond IP, ASN and FQDN, including both classification.identifier and
classification.type and more, but the classification.type alone helps us
most to get useful notifications at the moment.