Initial Version of a Configurable Shadowserver Parser
This parser contains of 2 Files:
parser.py and config.py

Config.py contains the definition of the feeds,
parser.py contains the logic.

The parser is capable of using conversion-functions those should
be defined in config.py. See the "UTC" function to get the idea.

This parser is intended to obsolete the other shadowserver parsers.