Add a fallback rule for notification lookup by ASN only
The design of the lookup by more complex rules is not clear enough yet
and at the moment it's useful to have a simple mechanism that lets us
use the data that can be imported with the current version of the
ripe_import.py. That data is ASN-only and has no particular rules for
selecting recipients based on other criteria such as
classification.type.

So this commit tries to implement a simple solution that lets us use
this data: if the lookup as it was implemented so far does not produce a
result, just lookup the contact for the events ASN if any and leave
selection of the template and data format to later stages of the
processing pipeline, such as intelmq-mailgen. The notification
description produced in this case has an empty template name and the
format "feed_specific". That format name is quite arbitrary but reflects
the way of template and format selection that seems most likely at the
moment, i.e. one based on the feed name.