shadowserver parser: Use csv writer in recover_line().
This fixes #600, by using the standard scv writer correct quoting is
guaranteed.  The faulty function conv_csv_shadowserver() has been
removed.

Please note, that the result of recover_line() will no longer try to
imitate the original format used in the shadowserver feeds.  But as
this was based on guessing and not on the original input this isn't a
loss of information.