Use an unlimited type for the raw field in the events table
This is a somewhat ad-hoc fix for the problem that the raw field can be
quite large and may not fit into the 2000 characters that are assumed if
no length is specified for a Base64 field in harmonization.conf. There's
probably a better solution but for now we only need it for the raw field
and that's the only field that uses the Base64 type.

See issue #547