Allow modify bot replacement rules to access regex match results.
The replacement rules for the modify bot can now access the results of
the regex matches that were performed when the event was matched to the
rules. This allows for more complex replacements because the
replacements can refer to e.g. regex groups to extract parts of other
fields. See the change in morefeeds.conf for an example.

In this form the change is somewhat experimental, but it demonstrates
the principle as well as some of the complications like having to wrap
the match results to give them a dict-like interface that can be used
from the format replacement fields.

(cherry picked from commit c771701be72888d93a3d281eef5bd2a21ebc65fe)