Add a way to explicitly inhibit notifications to certbund-contact bot
The existing mechanisms based on the notification_interval being -1
could be used to implement some of the requirements but they are too
cumbersome for the actual requirements. E.g. for specifying that for a
given IP address no notifications should be sent for events having a
given classification.type it would be necessary to create a dummy
organisation entry together with a suitable template entry even though
neither organisation nor template are actually needed.

The new approach uses a new table, inhibition, to specify which
host/network/as no notifications should be generated for events matching
a given classification.identifier. This mechanism is most likely too
simple and mathicng events based on e.g. classification.taxonomy or
classification.type will also be needed.