Modify Bot: improved example and unit test.
Removed morefeeds.conf default.__default action because it was
interfering with other rules. This was the case because all sections
will be considered, but in undefined order.

For example when running the bitdefender unittest:
Depending on which section was executed first,
the value of classification.identifier was different, because malware.name
was already modified by the bitdefender rule or not when the
default.__default action was executed.