Add information about inhibitions to certbund extra field
Instead of simply not adding notification information when inhibition
rules apply, we now add a source_inhibited (or destination_inhibited)
field to the certbund extra field to indicate this. Currently the value
is an empty list but it may be used to record the details of the
inhibition reasons later.