Separate more of the RIPE data handling from the DB interaction
Having data sanitation and construction of secondary data structures
like mappings from e.g. organisations to ASNs which only depend on the
RIPE data as such interspersed with the database interaction makes it
harder to understand the code and to reuse this part of the code in
other scripts. So we disentangle this and move most of it to the
ripe_data module.