Fix check for omitted parameters that may false even if not None
When checking whether an optional parameter was given whose default
value is None and which may have non-None values that are considered
false, check that it is not None. Do not simply check whether it's true.

In this particular case, the parameter abuse_c_to_org of
sanitize_role_list may be given as an empty dictionary in which case the
semantics of the parameter clearly means that the returned list should
be empty because no entry in role_list will have a value for nic-hdl
that is a key in the dict. As implemented, though, am empty dict would
lead to no role_list entry being removed based on nic-hdl.

The same happens with the optional parameter org_to_asn of
sanitize_organisation_list.