Introduction of the Enabled-Switch for Collectors.
This introduces the enabled-switch for all Collector-Bots.
By doing so it fixes #715.

The Collectorbot was extended with an own start-function
which checks if the enabled paramter is present and was set
to a representation of False. To check the "representation of
False" the function to_bool was introduced in utils.py, which
maps well-known-text to booleans. It has a a fallbackmode parameter
which determines if the function returns true of false if no
mapping  was found.

As long as to_bool does not evaluate to False the bot will be started.