Use the full sha256 hash in the deduplicator bot. Fixes #658
The deduplicator bot now uses the full sha256 hash as keys in the cache
database. Before it would end up using only 64 bits derived from the
sha256 hash due to using Python's built-in hash function which limits
hash values to a platform int, i.e. 64 bits on 64 bit platforms. See
issue #658 for details.

In order to be able to use the sha256 hash, this commit extracts the
actual computation of the hash of a message from the __hash__ method
into the new method hash and calls this new method from __hash__ for
compatibility. The new method returns the full sha256 hash value.