Do not copy the message to compute the hash in the deduplicator
This improves the speed of the deduplicator substantially because, as it
turns out, it spent much of its time when actually processing events
copying the message just to be able to remove the keys that should be
ignored in the hash.

To avoid that copy, the message's hash method now takes an additional,
optional parameter with a set of keys that should be ignored and the
deduplicator uses that parameter instead of modifying a copy of the
message.

Fixes #613 (at least it improves speed substantially)