Make certbund_contact bot add contact data without applying rules
The certbund_contact expert bot now simply looks up which contact
information in the database matches the event in question without
selecting one in case there are many. Making that choice is left to
other bots downstream.

Support for inhibitions has been removed for now as well.

This is an incompatible change that starts the refactoring of the whole
contact/notification handling of the certbund_contact bot and mailgen.