modify expert: Lowercase all malware names in morefeeds config.
The morefeeds config is identical to our current customer setup and
used by our packaging to be the readily available config for the
modify bot.  As per customer request all malware names should be
normalized to lower case.  (Capitalized names would be prefered if it
would be possible to capitalize arbitrary matches in modify bot, but
this is currently not feasible.)