Use a more convenient interface for the notification rule scripts
The entry points in the scripts are now called with a context object
that provides access to the event and other parameters. The contact and
directive information is now also represented by objects instead of
plain dicts which makes them a bit easier to use and less error prone
because e.g. mistyped attribute names for directives lead to
AttributeErrors instead of silently ignored dictionary entries-

The new interface is also easier to extend in some ways. For instance,
more information provided to the scripts can simply be added to the
context object and does not require changes in the scripts.