Consistently use 'ip' to refer to ip-addresses
In some places 'ipaddress' was used instead of 'ip'. Now the name of the
variables and other identifiers is consistent with the naming of this
property in IntelMQ. In the contact and rule bots the ip was usually
taken from source.ip or destination.ip.

This does include one incompatible change: The matched_field list in the
contacts attached to events by expert.py will now also call the field
"ip" instead of "ipaddress".