Add example script that selects the most specific contacts
Most of the actual logic is in the rulesupport module. The script
demonstrates how to get it into the rule.

The rules to decide which contacts are the most specific should be
equivalent to that used in the contact expert bot before this branch
for the notification rework started.

Add some simple tests for this, too.