Add support for aggregation parameters to notification directives
There are two new attributes for this, one for extra values that must be
equal in two different directives and one that names event fields whose
values must be equal. The latter makes it easier to specify aggregation
by e.g. ASN by not having to specify the ASN value because it can easily
be taken from the event in question.