certbund-contact: prevent redundant indices on db initialization.
When deriving a new table with LIKE from another one existing indices
are recreated too.  We don't want to rely on this automatism, to get
known and sensible names for the indices.  So we create all indices
after creating the original table as well as any derived tables.

Further more the UNIQUE constraint implies the creation of an index on
that column, so we don't create any indices manually for such columns.

NOTE: we should have update scripts for making these changes to
existing databases.