Make lookup_contacts more abstract wrt. manual/automatic contacts
Instead of supplying the table suffix ("" or "_automatic") as a
parameter, lookup_contacts now has a parameter for a value of the new
enum type Managed to indicate which set of contacts to query. The
mapping to the table name suffix is done in lookup_contacts, so the
callers have know less about the implementation details of the database.