Rework how contact information is communicated to the rule expert
This is a pretty substantial change and the new representation is quite
different from the old one. The goal was to preserve as much of the
information read from the database as possible so that the the rule
expert bot can provide all of the information to its scripts. They can
now base their decisions on all available contact information related to
the event.

In order to get to a working state, this commit contains the necessary
changes to both of the bots and the example scripts and test cases.