Add support for CIDR, ASN and FQDN in the contact DB
For each of the tables network, autonomous_system and fqdn there's now a
correspondong _annotation table for annotations like those already
supported in organisation_annotation. The query in lookup_contacts is
updated to retrieve those too so that they're included in the contact
information added to the events by the contact expert bot.