Simplify the representation of annotations
Instead of distinguishing tag-annotations from inhibition annotations
where the former have a tag attribute and the latter a condition
attribute, all annotations now have a tag attribute and an optional
condition attribute which if missing defaults to "true". Inhibition
annotations are now simply annotations with the tag "inhibition".

This simplifies the representation and allows all annotations to have
conditions, not just inhibition annotations.