Add support for RIPE's inetnum and inet6num data
This is a bit trickier than one would expect, because we cannot simply
store a single inetnum entry as a single row in network_automatic
because the IPv4 addresses are specified as an address interval, not as
a network in CIDR notation. These intervals can easily be converted to a
list of CIDR values but that can lead to two intervals sharing some CIDR
values which must be stored as a single network_automatic row to avoid
violating the unique constraint on the address column.

ripe_diff.py is updated only as much as necessary to adapt to the
interface change in ripe_data.py. It does not compare the network data
yet.