Allow restriction of RIPE network addresses to a specific country
The ripe_import.py and ripe_diff.py tools now have an additional command
line parameter --restrict-to-country which can be used to specify a
country code. Only inetnum and inet6num entries which have a country
attribute equal to this value are actually read from the files. This
reduces the number of entries to be processed substantially.