PEP8: Avoid assigning lambda expressions to variables
In this particular case, I think assigning a lambda expression to a
local variable is actually easier to read than using a def like the pep8
tool suggests because the intent is to overwrite the value of an already
existing local variable, not defining a local function. To avoid the
pep8 warning this commit uses the lambda directly as the default value
of the parameter.