Avoid empty match objects when querying the contact DB
The subqueries for ASNs and FQDNs always produced exactly one match
object even though not match was actually found. This match object was
not associated with any organisation and would not contain any
annotations but could conceivably influence the rules of the rule expert
bot.

With this commit no such match object will be produced anymore.