Avoid using the same sequence for several tables in the contact DB
The _automatic variants of the various tables in the contact DB shared
the sequence used for the primary keys with the non-_automatic tables
due to the LIKE ... INCLUDING ALL used when creating them. This clause
copies the expression used for the default values and therefore uses the
same sequence.

Sharing the sequences is not really a problem in practice, AFAICT, but
it is unusualy and may make future maintenance harder because most
developers and admins would not expect this.