Allow rule expert bot to optionally delete raw contact data
In most cases the information that the contact-DB expert bot adds to an
event is not needed anymore after the rule expert bot has generated the
notification directives. Therefore it usually a good idea to delete it
so that it doesn't take up space in e.g. the event DB.

This commit adds a new parameter for the rule expert bot,
remove_contact_data, which when true (i.e. not an empty string) makes
the bot remove the raw contact data after it has generated the
directives. This parameter defaults to 'true'.