Introduce id-columns for the autonomous_system tables
So far the ASN was the primary key in the autonomous_system tables. For
autonomous_system_automatic this cannot really be used because only the
combination of ASN and import_source can possibly be unique, because
multiple import sources might provide information about the same ASN. In
order to keep both variants of the tables consistent to avoid
unenecessary complications when trying to write queries that work on
both we introduce ID columns for both.

This commit does remove the unique constraints on the ASNs for now.
They'll probably be added in some form again later.