Make import_source part of the UNIQUE constraints
For the automatically managed data in the contact DB the import source
has to be part of the unique constraint because FQDN, network-address
and ASN cannot be kept unique across different import sources. At least
not without making this more complicated in other ways, by e.g. allow
multiple import sources per ASN, etc.

So this commit changes the UNIQUE constraints on the *_automatic tables.
The manually managed tables keep the UNIQUE constraints on the FQDN,
etc. Actually in the case of autonomous_system the UNIQUE is added again
after it was removed in a previous commit.