Remove printing of unattached contacts from ripe_diff.py
Most of the actual code had already been removed when the role tables
were removed because with the contacts directly referencing the
organisation they belong to and that organisation being mandatory there
cannot be unattached contacts in the database anymore. This commit
removes the rest of the code related to this.