Remove manual-contacts-from-file.py
The current version doesn't work anymore with the recent changes to the
contact DB and the script does not seem to be needed anymore. The
web-frontend will be more useful soo.

If it does turn out that manual-contacts-from-file.py is needed after
all, we can restore it and update it later.