Move conversion of networks addresses to ripe_data.py
The conversion from the string representation in the RIPE data to python
network address objects is needed for both ripe_import.py and
ripe_diff.py, so move it to ripe_data which is already used by both.

The conversion turns the string representation into a list of network
address objects from the ipaddress module.