Adapt lookup_contacts query to recent changes in the contact DB
The previous version of the query relied on the assumption that there
could only be one match in the fqdn table. This hasn't been true for
fqdn_automatic since its UNIQUE constraint was changed to be for the
pair (fqdn, import_source). It got even worse when the UNIQUE constraint
on fqdn in the manually maintained fqdn table was also removed.

The new version copes with there being multiple matches in the fqdn
tables and assigns the associated organisation IDs to the correct
matches.