Turn get_orgs_for_match into a context method
This reduces code duplication and provides a convenience method that may
be useful for other rules as well.

There's one difference between get_orgs_for_match and the new method
organisations_for_match: The former returns a set and the latter a list.
This doesn't make a difference for the places where get_orgs_for_match
was called because in both cases the code later simply iterates over the
elements. A list should be fine since a single match object is not
associated twice with the same organisation.