Fix default condition of Annotation objects
The default condition matches any event. The value used for this in the
constructor was wrong, though, as True is not a value condition object
(it's valid as a JSON value from which a condition is created by
expr_from_json). Use None instead and check this in Annotation.matches.
Add a test-case for this.