Ensure reference consistency in the rule script context
The rule scripts may change the contact information stored in the
context by assigning to the matches and organisations attributes. A
typical use case is removing matches and organisations that should not
be considered when deciding which notifications to create. Removing
organisations from the list may leave dangling references in the match
objects and also in the dictionary internally by the context object for
the organsiation ID lookups.

In order to ensure consistency the matches and organisations attributs
are now implemented as properties with setters that call the new method
ensure_data_consistency that cleans up any inconsistency. See its
doc-string and the comment for details. For good measure the rule expert
bot also calls this method between scripts in case the setters don't
catch all modifications (e.g. modifying the organisations list in-place
would not be caught).

The doc-string of Context now explicitly allows assignment to matches
and organisations and also explicitly warns not to modify the lists
in-place.