non-blocking bot/net starting/stopping
Until now, when starting or stopping a bot, we had to wait 0.25 sec. Intelmqctl waited for the bot to load and checked if it loaded properly.
Problems:
* When having 100 bots, we have to wait 25 s only to start the botnet.
* 0.25 s sometimes isn't enough: The bot seemed to be started without problem... than it immediately crashed due to an error, but in intelmq-manager, we saw that it was running till browser reload. It is so frustrating!

I've rewritten the class so that we wait 0.5 s whether we are starting/stopping a single bot or a whole botnet. (There is a new keyword parameter "getstatus" that handles that.)  
This seems to work very well. I'd be glad if you add this to the code.