cycling due to an invalid message failure
The invalid lines should be ignored.
Now, the bot is dying due to ```intelmq.lib.exceptions.InvalidValue: invalid value 'http:///ls5/forum.php' (<class 'str'>) for key 'source.url'```
In the moment of exception, the rest of lines won't get parsed and the message is acknowledged (so that the error repeats in few minutes again and again).

travis