Show info about manual networks in ripe_diff.py
ripe_diff.py already shows the manually maintained contact information
associated with the ASNs of organisations that have changed. With this
commit the same information is shown for the network addresses of
organisations.

Since the purpose of the find_overlaid_asns_db function has changed --
it's not just about ASNs anymore -- it has been renamed to
find_overlaid_manual_entries.