Improve message about which information an organisation refers to
In the message "Info: this entry is responsible for: X", it was not
clear enough, whether the X refers to what the situation will be like
when the change described by the diff output will have been applied or
what it's like before the change. In fact, it can be either, depending
on whether the organisation in question is being added, changed or
removed. In the latter case it was the situation before the change and
the opposite in the other cases.

The new message makes this clearer saying either "was responsible" or
"will be responsible", as appropriate.