Remove matches from the context when no org exists.
When the context has been altered by a script, for instance
to remove an organisation, the match is updated. It happpens,
that a match does not contain any org in this case.

Neverthelesse the match still existed in the context, which resulted
in problems when performing operations like "most_specific_matches",
as the matches without orgs for IP's are still more specific then matches
with org for ASN.

This commit introduces code which removes these empty matches.

To do so is safe, as long as all code honoring "inhibitions" for these
matches was run before the alteration of the context took place.

Currently the script 05inhibitions.py is the first script that is run,
so everything is fine.