Extract import_source for automatically managed organisations
In the contact database, automatically managed data has a field for the
import source. This information can be useful when deciding which
notification directives should be generated, so it's useful to include
this in the data looked up in the contact database.

This commit only does this for the organisation data. The information
associated with that organisation will usually come from the same
source, so it should be enough for many applications, and it's quite
easy to implement. It's not too difficult to extend this to some of the
other pieces of information should the need arise, though.