Fix comment in the certbund_contact test suite
The test suite was corrected recently in commit 65f7cf1 which changed
the expected output for one of the assertions to match the actual output
of the expression under test. That change was correct because the
behavior of the Context.ensure_data_consistency method was changed some
time earlier in commit 0ef048ad so that it does not only remove the
organisation reference from the match objects but also removes match
objects that do not reference any organisation anymore. So the change in
the actual output is indeed correct.

This commit completes that earlier correction by also updating the
comment that describes what the assertion tries to test.