Catch all exceptions, not just psycopg2 ones
The bare except that was converted to an "except psycopg2.Error:" with
commit 0bafcaf6 as a semantic change buried under many stylistic changes
suggested by pycodestyle really was intended to catch all exceptions,
not just the psycopg2 ones. Since using bare excepts is controversial
and often wrong---there's a reason pycodestyle warns about it, after
all---we catch BaseException instead which has the same semantics as a
bare except, but is less controversial.