Make getting the event's contact data more robust.
Do not rely on the contact information being present in the event at all
or for all sections. If the desired contact information is not found in
the event, get_certbund_contacts now returns empty contact information
that can actually be used by the rule expert bot and others. Previously
it would return an empty list, which attempted the same, but was
incorrect. Now it returns a dictionary with the expected keys which are
associated with the empty lists.