certbund-contact: make sections configurable
The contact bot can lookup contact information for both source and
destination event attributes and the rule expert bot can work with both
sets of contact information. Usually one is only interested in contact
information for the source, so we add a configuration option that can be
used to select one or both of "source" and "destination", with the
default being just "source". This can be done in both the contact expert
bot and the rule expert bot.