Remove unneeded test.skip_database() annotation
The TestCERTBundKontaktMockDBExpertBot class mocks the DB access
precisely so that it doesn't need an actual database connection, so that
the test.skip_database() annotation is not needed at all. In fact, that
annotation was harmful because it hid the problems this test case
actually has by telling the test suite runner not to run this test in
most circumstances.

This commit effectively reverts commit 7993c321cb5e