Add email_status support to the certbund bots
The contact bot now looks up all email addresses it retrieves from the
database in the new email_status table and adds that information to the
contact data it adds to the event. The rule expert bot now supports that
information in the contact data it presents to the rule scripts, too.