Fix country code restriction in ripe importer
The ripe importer can restrict the import of the CIDR information to a
given country code. The way this was implemented was incorrect in
several ways. The country code was assumed to be upper case only (in the
actual RIPE data it's sometimes lower case). The code assumed that there
was only one country code for each entry (there can be multiple). The
code also assumed that the value of the country:-Field was just the
country code (in fact it can contain something like a comment).

To handle all this correctly, the country code data is now normalized on
import (The comment is stripped and the result is converted to upper
case) and the check whether an entry belongs to a given country now
deals with multiple country values.

See the comments for some details.