Move conversion to network objects to ripe_data.py
In both places where the result of load_ripe_files is used,
convert_inetnum_to_networks and convert_inet6num_to_networks are called
to convert the network information to lists of Python network objects.
It's simpler to do that once in load_ripe_files.

It will also be needed in load_ripe_files because we want to process the
network related entries in order to determine contact data even for
entries without usable contact data by comparing networks.