Allow multiple iterations over a RIPE inetnum entry's CIDR list
When reading IPv4 inetnums, the address range is given as an interval of
IP addresses which we convert to the equivalent collection of CIDRs
using ipaddress.summarize_address_range. That function returns an
iterator, though, so it's not possible to use its result twice which we
may want to do. So we convert it to a list explicitly.