Handle unknown organisations in points_to_same_abuse_mailbox
The organisation referenced by the object passed to
points_to_same_abuse_mailbox may not actually be known. Treat it as if
the addresses were different, i.e. the function returns False in that
case. This should lead to the correct behaviour in the caller which will
then use the abuse-c address of the object in question.

It seems a bit strange that we did not run into this earlier, but
apparently the RIPE data associated with DE does not have any objects
with unknown organisations.