Postpone restriction of RIPE data by country
We want to be able to do some processing of the inetnum and inet6num
objects across all the data even when restricting the data written to
the database to networks that we know are in a particular country.
Therefore we cannot restrict the data when reading it. We have to do it
a bit later.