Consider enclosing networks when searching contact data for networks
Some RIPE entries for inetnum or inet6num do not have usable contact
information but entries for supernets of these entries might have. This
is a problem in particular when restricting the information to a single
country because sometimes the supernets are associated with other
countries and therefore would not end up in the contact database.

See the comments and doc-strings for details.