Initialize logging_level attribute
The logging level attribute is used in the `check` method (an
passed on to the `setup_list_logging` method). Usually the check
method is only run using the `intelmqctl` command, which runs the
`run` method first which initializes the attribute. If we want the
`check` method to also be usable without having to run the `run`
method before, the `logging_level` attribute has to at least exist.